Abstract out and use RampState hook to manage state that exists on the RampSmartContract
