+++ client/src/contexts/Ramp/index.ts
@@ -1,0 +1,2 @@
+export { default as RampContext } from './RampContext'
+export { default as RampProvider } from './RampProvider'
