+++ client/src/contexts/RampRegistration/RampRegistrationContext.ts
@@ -7,7 +7,6 @@ interface RampRegistrationValues {
   isRegistered: boolean;
   registrationHash: string;
   registeredVenmoId: string;
-  minimumDepositAmount: number;
   deposits: Deposit[];
 }
 
@@ -15,7 +14,6 @@ const defaultValues: RampRegistrationValues = {
   isRegistered: false,
   registrationHash: '',
   registeredVenmoId: '',
-  minimumDepositAmount: 0,
   deposits: [],
 };
