+++ client/src/hooks/useRampState.ts
@@ -1,0 +1,9 @@
+import { useContext } from 'react'
+
+import { RampContext } from '../contexts/Ramp'
+
+const useRampState = () => {
+  return { ...useContext(RampContext) }
+}
+
+export default useRampState
